--- Team Project 5/Assets/Script/Hikaru_Yamada/BarOpen.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1696a8d67d6db6e4badf3d4b1e663cba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
